Update script.js
Hi mate, you don't want to use var anymore. If the variable is not going to be mutated, use const. If it is going to mutated, use let. Read up more about the benefits of using const and let over var. 
e.g. This article on why using let is better than using var: https://stackoverflow.com/questions/762011/whats-the-difference-between-using-let-and-var
Also, you had an unnecessary if statement inside the calcProductChangeRate function.